+++ config/l10n.php
@@ -1,0 +1,29 @@
+<?php
+
+return [
+
+    /*
+    |--------------------------------------------------------------------------
+    | Hide default locale
+    |--------------------------------------------------------------------------
+    |
+    | Determines whether to omit the locale segment from URLs for
+    | the fallback language. E.g., forces /example instead of /en/example.
+    |
+    */
+
+    'hide_default_locale' => true,
+
+    /*
+    |--------------------------------------------------------------------------
+    | Hide alias locale
+    |--------------------------------------------------------------------------
+    |
+    | Hides the locale prefix for defined aliases. E.g., ['it' => 'esempio]
+    | generates /esempio instead of /it/esempio.
+    |
+    */
+
+    'hide_alias_locale' => false,
+
+];